requirements.txt
torch==2.1.0
packaging
fschat
peft
transformers==4.35.0
pytz
omegaconf
hydra-core
packaging
peft==0.6.1
deepspeed==0.12.2
sentencepiece
wandb
rich
pandas<2.0.0
easydict
bidict
scikit-learn
openai
einops
chardet
bitsandbytes
